Add docker dev environment
The other one should be clean and exactly like the build environment. The dev one is built on the python image and should stick around--don't call docker run with --rm on it.
Also added the run-dev-env.ps1 script to kick it off on windows. It uses a bind-style mount so that you can change code in windows and test it in the container without restarting.