Fix make test step to exit with a non-zero code if tests fail
By default pipes in bash return the exit code of the last thing in the pipe.
In our case we were piping to tee. It turns out that the output of
unittest doesn't use stdout, so writes to the console directly anyway and tee isn't needed.

However, had to kick unittest from another python script to force the return code to be nonzero if tests failed.